LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/uievents/click/click_event_target_siblings-expected.txt

Click targeting when targets of down and up are sibling elements

This test verifies that click event always goes to the first common ancestor of down and up event targets.
Press down the primary button on red div and move to blue box and release.
Press down the primary button on button b1 and move to button b2 and release.
Press down the primary button on input i1 and move to input i2 and release.
Press down the primary button on link 1 and move to link 2 and release.
Click done.
b1  b2

link1
link2
Done

FAIL Click targets the nearest common ancestor assert_equals: Click should be sent to the nearest common ancestor expected "mousedown@red_div,mouseup@blue_div,click@div_container,mousedown@button1,mouseup@button2,click@button_container,mousedown@link1,mouseup@link2,click@link_container" but got "mousedown@button1,mouseup@button1,click@button1"
